debugging

+++ include/tuple.cuh
@@ -8,12 +8,12 @@ using u64 = unsigned long long;
 using u32 = unsigned long;
 
 #ifndef USE_64_BIT_TUPLE
-using column_type = u32;
+using column_type = uint32_t;
 #else
-using column_type = u64;
+using column_type = uint64_t;
 #endif
 using tuple_type = column_type *;
-using tuple_size_t = u32;
+using tuple_size_t = unsigned long long;
 
 #define EMPTY_HASH_ENTRY UINT64_MAX
 #define MAX_ARITY 10
